Simplify userProile to use target size
Rather than scale the canvas to a scale, use a specific width and height.
This has the benifit of being agnostic to the previous canvas size, which
had caused issues previosuly when the image was already small, then was
either scaled to a tiny or massive size.

This also allows user to go back and change their image after setting it
if they click back, in response to tom's comment during gameplay